refactor: changed cards wrapper layout, separeted vehicles pages by folder
