+++ src/components/Button/styles.module.scss
@@ -11,6 +11,7 @@
     background-color: #65DCC7;
     border: 0;
     border-radius: 100px;
+    box-shadow: 1px 2px 5px 0px rgba(0,0,0,0.2);
     @include laptop{
         max-width: 365px;
         margin: 0 auto;
