+++ src/components/Card/Card.module.scss
@@ -18,6 +18,12 @@
   h2 {
     margin: 0 0 1rem 0;
     font-size: 1.5rem;
+    color:#fff;
+  }
+  .content{
+    .info{
+      color:#fff
+    }
   }
 
   .actions{
