+++ src/components/Inputs/Search/style.module.scss
@@ -6,7 +6,7 @@
   padding: 10px 30px;
   background-color: rgba(101, 220, 199, 0.3);
   border-radius: 100px;
-  box-shadow: 1px 2px 5px 0px rgba(0,0,0,0.3);
+  box-shadow: 1px 2px 5px 0px rgba(0,0,0,0.2);
 }
 .search_icon {
   display: flex;
