+++ src/pages/Vehicles/Create/Vehicles.module.scss
@@ -1,0 +1,35 @@
+@import '../../../styles/break-points.scss';
+
+.Vehicles {
+  display: grid;
+  padding: 1rem 2rem;
+
+  .main {
+    display: grid;
+    gap: 1rem;
+    width: 100%;
+  }
+  .section_title {
+    font-size: 18px;
+  }
+  .search_filter_box {
+    display: grid;
+    grid-template-columns: 80% 20%;
+
+    .filter_button {
+      background-color: transparent;
+      border: 0;
+    }
+  }
+  .cards_wrapper {
+    display: grid;
+    grid-template-columns: 1fr;
+    gap: 10px;
+    @include tablet {
+      grid-template-columns: 1fr 1fr;
+    }
+    @include laptop {
+      grid-template-columns: 1fr 1fr 1fr;
+    }
+  }
+}
