+++ src/pages/Vehicles/Create/index.tsx
@@ -1,0 +1,15 @@
+
+import styles from './Vehicles.module.scss'
+
+const CreateVehicle = () => {
+
+
+  return (
+    <div className={styles.Vehicles}>
+      <main className={styles.main}>
+      </main>
+    </div>
+  )
+}
+
+export default CreateVehicle
