--- src/pages/Vehicles/Vehicles.module.scss
@@ -1,36 +1,0 @@
-@import '../../styles/break-points.scss';
-
-.Vehicles {
-  display: grid;
-  padding: 1rem 2rem;
-
-  .main {
-    display: grid;
-    gap: 1rem;
-    width: 100%;
-  }
-  .search_filter_box{
-    display: grid;
-    grid-template-columns: 80% 20%;
-
-    .filter_button{
-      background-color: transparent;
-      border: 0;
-    }
-  }
-  @include tablet {
-
-  }
-
-  @include laptop {
-
-  }
-
-  @include desktop {
-
-  }
-
-  @include desktop-large {
-
-  }
-}
