+++ src/pages/Vehicles/Vehicles.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import VehiclesPage from './index';
+import VehiclesPage from './Home/index';
 
 test('renders learn react link', () => {
   render(<VehiclesPage />);
